requirements.txt
Pillow>=9.0.0
pyparsing>=3.0.0
